+++ projects/rx/src/test/kotlin/examples/promises.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015 Mark Platvoet<mplatvoet@gmail.com>
+ * Copyright (c) 2016 Mark Platvoet<mplatvoet@gmail.com>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -16,13 +16,20 @@
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
  * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
 
-project.description = "Progress tracking Kovenant extensions"
+package nl.komponents.kovenant.rx.examples
 
-dependencies {
-    compile project(':kovenant-core')
-    compile 'nl.komponents.progress:progress-core:1.0.0'
+import nl.komponents.kovenant.rx.toObservable
+import nl.komponents.kovenant.task
+import rx.schedulers.Schedulers
+
+fun main(args: Array<String>) {
+    val observable = task { 42 }.toObservable()
+
+    observable.subscribeOn(Schedulers.computation()).subscribe {
+        println(it)
+    }
 }
+